feat: compliance SRS v3.11 + build ESP32-S3 N16R8 com PSRAM
- PA-C01 a PA-C09 implementados (ALMs, CORS, salt, SAFE_OFF,
  atomic_write, alerts, import/export, topbar)
- AlmMonitor com 65 ALMs em alm_monitor.c/.h
- Web API: CORS headers, /api/v1/state, /api/v1/export, /api/v1/import
- Auth: hash salgado (SHA-256 + salt 32B em NVS), rate limit, logout
- SAFE_OFF: reason/timestamp/source ALM + relay_all_off + EventLog
- Storage: fflush + fsync(fileno()) antes de rename (escrita atomica)
- UI: topbar com state_badge, sd_label, selftest_label, alert_count
- Correcao compilacao: includes paths, const qualifier, macros duplicadas,
  LVGL read_cb (bool->void), NVS iterator API v5.3, wifi_ctl_mode_t
- Build: root CMakeLists.txt, scripts build.ps1/setup.ps1
- Target: ESP32-S3 N16R8, flash 16MB, custom partitions (OTA + spiffs),
  PSRAM 8MB octal 80MHz, LVGL fonts 10/12/14/16/18/28/48

+++ firmware/services/conf_ctl.c
@@ -3,6 +3,7 @@
 
 #include <stdio.h>
 #include <string.h>
+#include <inttypes.h>
 
 #include "esp_log.h"
 
@@ -68,9 +69,9 @@ esp_err_t conf_ctl_export_json(char *buf, size_t buf_size)
 {
     if (!buf || buf_size == 0) return ESP_ERR_INVALID_ARG;
     snprintf(buf, buf_size,
-        "{\"temp_min\":%.1f,\"temp_max\":%.1f,\"brightness\":%u,\"sleep_timeout\":%u,\"timezone\":\"%s\"}",
+        "{\"temp_min\":%.1f,\"temp_max\":%.1f,\"brightness\":%"PRIu32",\"sleep_timeout\":%"PRIu32",\"timezone\":\"%s\"}",
         s_saved.sensor.temp_min_c, s_saved.sensor.temp_max_c,
-        s_saved.display.brightness, s_saved.display.sleep_timeout_s,
+        (uint32_t)s_saved.display.brightness, (uint32_t)s_saved.display.sleep_timeout_s,
         s_saved.datetime.timezone);
     return ESP_OK;
 }
